refactor(rag): consolidate RAGComponentFactory and its implementation into a single file
